+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.16.0+45 (in progress)
 
 - add search and sort options in library -> albums, library -> artists, library -> playlists.
+- add "save to files" option for songs in music player page.
 - techincal changes
   - bug fix in fetching album art.
   - update test cases.